+++ core/model/user.go
@@ -1,7 +1,7 @@
 package model
 
 import (
-	"github.com/rokwire/core-auth-library-go/v2/tokenauth"
+	"github.com/rokwire/rokwire-building-block-sdk-go/services/core/auth/tokenauth"
 )
 
 // User auth wrapper