+++ driver/web/rest/apis.go
@@ -383,7 +383,7 @@ func (h ApisHandler) GetPollEvents(user *model.User, w http.ResponseWriter, r *h
 			if err != nil {
 				log.Printf("Error on apis.GetPollEvents(): %s", err)
 			}
-			log.Printf(string(jsonData))
+			log.Print(jsonData)
 			w.Write(jsonData)
 			flusher.Flush()
 		} else {